requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0
jieba>=0.42.1
scikit-learn>=1.3.0
numpy>=1.26.0
PyMySQL>=1.1.0
websocket-client>=1.6.0
